\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,prev-method}

\title{Methods for Function \code{show} in Package \pkg{prevalence}}

\description{Show objects of class \code{prev}}

\usage{\S4method{show}{prev}(object)}

\arguments{
  \item{object}{An object of class \code{prev}}
}

\section{Methods}{
  \describe{
    \item{\code{signature(object = "prev")}}{
      Corresponds to \code{print(object)}
    }
  }
}

\seealso{
  \code{\link{prev-class}}
}

\keyword{methods}
